Reorganizes object classes
