+++ assign2/src/model/question/actions.ts
@@ -1,26 +1,26 @@
-import User from "../User";
-import Tag from "../Tag";
-import Question from "../Question";
+import User from "../objects/User";
+import Tag from "../objects/Tag";
+import Question from "../objects/Question";
 import {FILTER_BY_TAG, NEW_POST, QuestionActions, QuestionsState, SEARCH_BY_TITLE} from "./types";
 
 
 
 
-function doNewPost(title: string, text: string, author: User): QuestionActions{
+export function doNewPost(title: string, text: string, author: User): QuestionActions{
     return {
         type: NEW_POST,
         newPost: { title: title, text: text, author: author }
     };
 }
 
-function doFilterByTag(tag: Tag): QuestionActions{
+export function doFilterByTag(tag: Tag): QuestionActions{
     return {
         type: FILTER_BY_TAG,
         tag
     };
 }
 
-function doSearchByTitle(searchedText: string): QuestionActions{
+export function doSearchByTitle(searchedText: string): QuestionActions{
     return {
         type: SEARCH_BY_TITLE,
         searchedText
